+++ dnssd/unicastresolver_test.go
@@ -34,7 +34,7 @@ var _ = Context("UnicastResolver", func() {
 			TargetPort: 12345,
 			Priority:   10,
 			Weight:     20,
-			Attributes: []Attributes{
+			Attributes: AttributeCollection{
 				NewAttributes().
 					WithPair("<key>", []byte("<instance-a>")),
 			},
@@ -51,7 +51,7 @@ var _ = Context("UnicastResolver", func() {
 			TargetPort: 12345,
 			Priority:   10,
 			Weight:     20,
-			Attributes: []Attributes{
+			Attributes: AttributeCollection{
 				NewAttributes().
 					WithPair("<key>", []byte("<instance-b0>")),
 				NewAttributes().