+++ CHANGELOG.md
@@ -10,6 +10,20 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Added `dnssd.AttributeCollection` type
+
+### Changed
+
+- **[BC]** Changed `dnssd.ServiceInstance.Attributes` type from `[]Attributes` to `AttributeCollection`
+
+### Removed
+
+- **[BC]** Removed `dnssd.AttributeCollectionsEqual()` function
+
 ## [0.3.1] - 2023-08-15
 
 ### Changed